TweetPostRequest.java class file is created for adding TweetPostRequest class which was initialy inside TwitterRunner.java which is deleted
