+++ src/main/java/com/resource/TweetPostRequest.java
@@ -1,0 +1,15 @@
+package com.resource;
+public class TweetPostRequest
+{
+    String message;
+    TweetPostRequest()
+    {
+    }
+    public String getMessage()
+    {
+        return message;
+    }
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
